backend/requirements.txt
django
channels
daphne
django-cors-headers
djangorestframework
python-dotenv
google-generativeai